update and refactor tests
